% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd10am.pph-package.R
\docType{package}
\name{icd10am.pph-package}
\alias{icd10am.pph}
\alias{icd10am.pph-package}
\title{icd10am.pph: Potentially Preventable Hospitalisations for ICD-10-AM Coded Data}
\description{
Identifies Potentially Preventable Hospitalisations (PPH) in Australian admitted patient care data coded using ICD-10-AM, using the AIHW national PPH indicator specification.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/healthpolicy/icd10am.pph}
  \item Report bugs at \url{https://github.com/healthpolicy/icd10am.pph/issues}
}

}
\author{
\strong{Maintainer}: Jim Pearse \email{ejpearse@gmail.com}

}
\keyword{internal}
